+++ chainerx_cc/chainerx/cuda/cuda_device/copy.cu
@@ -12,8 +12,8 @@
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
 #include "chainerx/kernels/creation.h"
+#include "chainerx/kernels/misc.h"
 #include "chainerx/routines/creation.h"
-#include "chainerx/routines/misc.h"
 
 namespace chainerx {
 namespace cuda {
