+++ chainerx_cc/chainerx/cuda/cuda_device/dot.cu
@@ -25,10 +25,10 @@
 #include "chainerx/kernels/creation.h"
 #include "chainerx/kernels/linalg.h"
 #include "chainerx/kernels/math.h"
+#include "chainerx/kernels/misc.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
-#include "chainerx/routines/misc.h"
 
 namespace chainerx {
 namespace cuda {
