+++ chainerx_cc/chainerx/routines/CMakeLists.txt
@@ -21,7 +21,6 @@ install(FILES
     logic.h
     manipulation.h
     math.h
-    misc.h
     normalization.h
     pooling.h
     routines_util.h
