+++ chainerx_cc/chainerx/routines/creation.cc
@@ -19,8 +19,8 @@
 #include "chainerx/dtype.h"
 #include "chainerx/graph.h"
 #include "chainerx/kernels/creation.h"
+#include "chainerx/kernels/misc.h"
 #include "chainerx/macro.h"
-#include "chainerx/routines/misc.h"
 #include "chainerx/routines/type_util.h"
 #include "chainerx/scalar.h"
 #include "chainerx/shape.h"
